requirements.txt
click==8.0.3
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
PyYAML==6.0
